django/system/requirements/main.txt
Django>=4.2.0
psycopg2-binary>=2.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
flake8>=6.0.0
isort>=5.0.0
black==24.10.0